// src/app/components/shared/models/order.ts
export class InformationToForm {
  reversionRequestStatusId: number;
}

/**
 * Entidad que se emplea en los formularios de las diferente tablas de la aplicación,
 * Permite indicar el titulo, el tipo de formulario y la información que se le desee pasar al formulario.
 * @export
 * @class SearchFormEntity
 */
export class SearchFormEntity {
  title: string;
  type_form: string;
  btn_title: string;
  title_for_search: string;
  information: InformationToForm;
}